[ty] Extract `ty_site_packages` crate
