+++ crates/ruff_graph/Cargo.toml
@@ -22,6 +22,7 @@ ruff_python_ast = { workspace = true }
 ruff_python_parser = { workspace = true }
 ty_module_resolver = { workspace = true }
 ty_python_semantic = { workspace = true }
+ty_site_packages = { workspace = true }
 
 anyhow = { workspace = true }
 clap = { workspace = true, optional = true }
