+++ crates/ty_python_semantic/src/program.rs
@@ -1,16 +1,12 @@
-use std::sync::Arc;
-
 use crate::Db;
 use crate::python_platform::PythonPlatform;
 
-use ruff_db::diagnostic::Span;
-use ruff_db::files::system_path_to_file;
-use ruff_db::system::{SystemPath, SystemPathBuf};
+use ruff_db::system::SystemPath;
 use ruff_python_ast::PythonVersion;
-use ruff_text_size::TextRange;
 use salsa::Durability;
 use salsa::Setter;
 use ty_module_resolver::SearchPaths;
+use ty_site_packages::PythonVersionWithSource;
 
 #[salsa::input(singleton, heap_size=ruff_memory_usage::heap_size)]
 pub struct Program {
@@ -90,74 +86,4 @@ pub struct ProgramSettings {
     pub python_version: PythonVersionWithSource,
     pub python_platform: PythonPlatform,
     pub search_paths: SearchPaths,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Default, get_size2::GetSize)]
-pub enum PythonVersionSource {
-    /// Value loaded from a project's configuration file.
-    ConfigFile(PythonVersionFileSource),
-
-    /// Value loaded from the `pyvenv.cfg` file of the virtual environment.
-    /// The virtual environment might have been configured, activated or inferred.
-    PyvenvCfgFile(PythonVersionFileSource),
-
-    /// Value inferred from the layout of the Python installation.
-    ///
-    /// This only ever applies on Unix. On Unix, the `site-packages` directory
-    /// will always be at `sys.prefix/lib/pythonX.Y/site-packages`,
-    /// so we can infer the Python version from the parent directory of `site-packages`.
-    InstallationDirectoryLayout { site_packages_parent_dir: Box<str> },
-
-    /// The value comes from a CLI argument, while it's left open if specified using a short argument,
-    /// long argument (`--extra-paths`) or `--config key=value`.
-    Cli,
-
-    /// The value comes from the user's editor,
-    /// while it's left open if specified as a setting
-    /// or if the value was auto-discovered by the editor
-    /// (e.g., the Python environment)
-    Editor,
-
-    /// We fell back to a default value because the value was not specified via the CLI or a config file.
-    #[default]
-    Default,
-}
-
-/// Information regarding the file and [`TextRange`] of the configuration
-/// from which we inferred the Python version.
-#[derive(Debug, PartialEq, Eq, Clone, get_size2::GetSize)]
-pub struct PythonVersionFileSource {
-    path: Arc<SystemPathBuf>,
-    range: Option<TextRange>,
-}
-
-impl PythonVersionFileSource {
-    pub fn new(path: Arc<SystemPathBuf>, range: Option<TextRange>) -> Self {
-        Self { path, range }
-    }
-
-    /// Attempt to resolve a [`Span`] that corresponds to the location of
-    /// the configuration setting that specified the Python version.
-    ///
-    /// Useful for subdiagnostics when informing the user
-    /// what the inferred Python version of their project is.
-    pub(crate) fn span(&self, db: &dyn Db) -> Option<Span> {
-        let file = system_path_to_file(db, &*self.path).ok()?;
-        Some(Span::from(file).with_optional_range(self.range))
-    }
-}
-
-#[derive(Eq, PartialEq, Debug, Clone, get_size2::GetSize)]
-pub struct PythonVersionWithSource {
-    pub version: PythonVersion,
-    pub source: PythonVersionSource,
-}
-
-impl Default for PythonVersionWithSource {
-    fn default() -> Self {
-        Self {
-            version: PythonVersion::latest_ty(),
-            source: PythonVersionSource::Default,
-        }
-    }
 }
