+++ crates/ty_site_packages/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "ty_site_packages"
+version = "0.0.0"
+publish = false
+authors = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+homepage = { workspace = true }
+documentation = { workspace = true }
+repository = { workspace = true }
+license = { workspace = true }
+
+[dependencies]
+ruff_annotate_snippets = { workspace = true }
+ruff_db = { workspace = true }
+ruff_python_ast = { workspace = true }
+ruff_python_trivia = { workspace = true }
+ruff_source_file = { workspace = true }
+ruff_text_size = { workspace = true }
+ty_static = { workspace = true }
+
+camino = { workspace = true }
+colored = { workspace = true }
+get-size2 = { workspace = true }
+indexmap = { workspace = true }
+strum = { workspace = true }
+strum_macros = { workspace = true }
+tracing = { workspace = true }
+
+[dev-dependencies]
+ruff_db = { workspace = true, features = ["testing", "os"] }
+
+[lints]
+workspace = true
